1) External triggrer flag added to timerClass, function added to    set/unset the flag. 2) Imagegrab thread timing estimates for expected time slightly    modified.

+++ src/plugin/jaaba_plugin/jaaba_plugin.hpp
@@ -251,7 +251,7 @@ namespace bias
             void write_score_final(std::string file, unsigned int numFrames,
                 vector<PredData>& pred_score);
             bool jaabaSkipFrame(uint64_t& ts_cur, uint64_t& ts_prev,
-                int& frameCount, uint64_t wait_thres);
+                unsigned long& frameCount, uint64_t wait_thres);
             void waitForEmptyHOGHOFAvgQueue(LockableQueue<vector<float>>& avg_que);
 
             //void writeAllFeatures(string filename, vector<float>& feat_out,
